refactor: split node proxy stream handling
